Move login_submission_legacy_convert, login_id_thirdparty_from_phone methods
We'll be using them in AuthHandler in the separate PR, so move them to AuthHandler
instead of importing them from LoginRestServlet
